# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.Test:
    properties:
      message:
        type: string
    type: object
  model.IdResponse:
    properties:
      id:
        type: string
    type: object
  model.TodoItem:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /ping:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Test'
      summary: ping example
      tags:
      - example
  /todos:
    get:
      consumes:
      - application/json
      description: list todo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TodoItem'
            type: array
        "400":
          description: Bad Request
      summary: list todo
      tags:
      - example
    post:
      consumes:
      - application/json
      description: create todo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IdResponse'
        "400":
          description: Bad Request
      summary: create todo
      tags:
      - example
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: delete todo
      parameters:
      - description: todo id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: delete todo
      tags:
      - example
    get:
      consumes:
      - application/json
      description: show detail
      parameters:
      - description: todo id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TodoItem'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: show detail
      tags:
      - example
    put:
      consumes:
      - application/json
      description: update todo
      parameters:
      - description: todo id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: update todo
      tags:
      - example
swagger: "2.0"